% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateXts.R
\name{updateXts}
\alias{updateXts}
\alias{update.xts}
\title{updateXts}
\usage{
updateXts(object, new_data)
}
\arguments{
\item{object}{old xts object}

\item{new_data}{an xts object with new data}
}
\value{
an updated xts object
}
\description{
use new data to update an xts object.
}
\examples{
\dontrun{
library(xts)
a = as.xts(1:10, order.by = seq(Sys.Date() - 15, Sys.Date()-6, "days"))
a

b = as.xts(101:110, order.by = seq(Sys.Date() - 9, Sys.Date(), "days"))
b

updateXts(a,b)
update(a,b)
}
}
