% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateXts.R
\name{updateXts}
\alias{updateXts}
\title{updateXts}
\usage{
updateXts(x, new_data)
}
\arguments{
\item{x}{old xts object}

\item{new_data}{new data}
}
\value{
a updated xts object
}
\description{
use new data to update the old xts object(x)
}
\examples{



}
